requirement.txt
audioread==3.0.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
crayons==0.4.0
decorator==5.1.1
Faker==33.3.1
idna==3.10
imageio==2.36.1
joblib==1.4.2
kagglehub==0.3.6
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
msgpack==1.1.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
piexif==1.1.3
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
pycparser==2.22
pydub==0.25.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.8.0
requests==2.32.3
scikit-image==0.25.0
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
soundfile==0.13.0
soxr==0.5.0.post1
stegano==0.11.4
threadpoolctl==3.5.0
tifffile==2025.1.10
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
